requirements.txt
Django==3.1.12
django-treebeard==4.4
django-cms==3.8.0
django-filer==2.0.2
django-cleanup==5.2.0
djangocms-text-ckeditor==4.0.0
djangocms-picture==3.0.0
djangocms-file==3.0.0
djangocms-link==3.0.0
djangocms-style==3.0.0
djangocms-video==3.0.0
djangocms-icon==2.0.0
djangocms-googlemap==2.0.0
djangocms-bootstrap4==2.0.0
